tests/wpt/mozilla/tests/mozilla/window_performance_topLevelDomComplete.html

<!doctype html>
<meta charset="utf-8">
<title>Performance topLevelDomComplete</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body>
<script>
async_test(function (t) {
  var observer = new PerformanceObserver(
    window.onload = t.step_func(function (entryList, obs) {

      var entries = window.performance.getEntries();
      assert_greater_than(entries.length, 0);
      var navigation = entries[0];
      assert_true(navigation instanceof PerformanceNavigationTiming);

      assert_true(navigation.domContentLoadedEventStart <= navigation.topLevelDomComplete);
      assert_true(navigation.topLevelDomComplete <= navigation.domComplete);

      observer.disconnect();
      t.done();
    }));

  observer.observe({entryTypes: ["navigation"]});

}, "top level dom complete");
</script>
</body>
